refactor(action-center): extract the row so one guard covers both surfaces (#433)
The panel and the coming full page must not each own a copy of "how an
action item becomes a link". Two copies drift, and the way they drift is
that one of them forgets to run safeDeepLink and starts rendering rows
that go nowhere. There is now one row component; it takes an href that
has already been proved to resolve, and neither surface can render an
item any other way.

The date helpers move to their own module rather than sitting beside a
component: a file exporting both breaks Vite fast refresh, which the
react-refresh rule enforces at error level.
